fix(1234-Windows-port-new): fix s3 handler stol issue
* fix(1234-Windows-port-new): fix windows port.

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix FTP passive port

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix dynamodb container creation

* fix(1234-Windows-port-new): fix application startup

* Merge branch 'main' into fix/1234-Windows-port-new

* fix(1234-Windows-port-new): fix application startup

* fix(1234-Windows-port-new): fix s3 handler stol issue

* Merge branch 'main' into fix/1234-Windows-port-new

+++ src/service/src/s3/S3Handler.cpp
@@ -355,7 +355,7 @@ namespace AwsMock::Service {
                     Dto::S3::UploadPartCopyRequest s3Request;
                     s3Request.region = region;
                     s3Request.uploadId = uploadId;
-                    s3Request.partNumber = std::stoi(partNumber);
+                    s3Request.partNumber = std::stol(partNumber);
                     s3Request.targetBucket = clientCommand.bucket;
                     s3Request.targetKey = clientCommand.key;
